txn: move some methods in `session_txn` to internal pacakge
