+++ sessiontxn/staleread/provider.go
@@ -26,6 +26,7 @@ import (
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/sessiontxn"
+	"github.com/pingcap/tidb/sessiontxn/internal"
 	"github.com/pingcap/tidb/table/temptable"
 )
 
@@ -89,7 +90,7 @@ func (p *StalenessTxnContextProvider) OnInitialize(ctx context.Context, tp sessi
 // with the staleness snapshot ts. After that, it sets the relevant context variables.
 func (p *StalenessTxnContextProvider) activateStaleTxn() error {
 	var err error
-	if err = sessiontxn.CommitBeforeEnterNewTxn(p.ctx, p.sctx); err != nil {
+	if err = internal.CommitBeforeEnterNewTxn(p.ctx, p.sctx); err != nil {
 		return err
 	}
 
@@ -108,7 +109,7 @@ func (p *StalenessTxnContextProvider) activateStaleTxn() error {
 	txn.SetVars(sessVars.KVVars)
 	txn.SetOption(kv.IsStalenessReadOnly, true)
 	txn.SetOption(kv.TxnScope, txnScope)
-	sessiontxn.SetTxnAssertionLevel(txn, sessVars.AssertionLevel)
+	internal.SetTxnAssertionLevel(txn, sessVars.AssertionLevel)
 	is, err := GetSessionSnapshotInfoSchema(p.sctx, p.ts)
 	if err != nil {
 		return errors.Trace(err)
@@ -208,7 +209,7 @@ func (p *StalenessTxnContextProvider) GetSnapshotWithStmtReadTS() (kv.Snapshot, 
 	}
 
 	sessVars := p.sctx.GetSessionVars()
-	snapshot := sessiontxn.GetSnapshotWithTS(p.sctx, p.ts)
+	snapshot := internal.GetSnapshotWithTS(p.sctx, p.ts)
 
 	replicaReadType := sessVars.GetReplicaRead()
 	if replicaReadType.IsFollowerRead() {
