+++ sessiontxn/txn_manager_test.go
@@ -23,10 +23,14 @@ import (
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/ast"
+	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessiontxn"
+	"github.com/pingcap/tidb/sessiontxn/internal"
 	"github.com/pingcap/tidb/sessiontxn/staleread"
+	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/testkit"
+	"github.com/pingcap/tidb/tests/realtikvtest"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/oracle"
 )
@@ -306,7 +310,7 @@ func TestGetSnapshot(t *testing.T) {
 			check: func(t *testing.T, sctx sessionctx.Context) {
 				ts, err := mgr.GetStmtReadTS()
 				require.NoError(t, err)
-				compareSnap := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err := mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap, snap))
@@ -316,7 +320,7 @@ func TestGetSnapshot(t *testing.T) {
 				tk.MustQuery("select * from t for update").Check(testkit.Rows("1", "3", "10"))
 				ts, err = mgr.GetStmtForUpdateTS()
 				require.NoError(t, err)
-				compareSnap2 := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap2 := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err = mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.False(t, isSnapshotEqual(t, compareSnap2, snap))
@@ -336,7 +340,7 @@ func TestGetSnapshot(t *testing.T) {
 			check: func(t *testing.T, sctx sessionctx.Context) {
 				ts, err := mgr.GetStmtReadTS()
 				require.NoError(t, err)
-				compareSnap := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err := mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap, snap))
@@ -346,7 +350,7 @@ func TestGetSnapshot(t *testing.T) {
 				tk.MustQuery("select * from t").Check(testkit.Rows("1", "3", "10"))
 				ts, err = mgr.GetStmtForUpdateTS()
 				require.NoError(t, err)
-				compareSnap2 := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap2 := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err = mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap2, snap))
@@ -365,7 +369,7 @@ func TestGetSnapshot(t *testing.T) {
 			check: func(t *testing.T, sctx sessionctx.Context) {
 				ts, err := mgr.GetStmtReadTS()
 				require.NoError(t, err)
-				compareSnap := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err := mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap, snap))
@@ -375,7 +379,7 @@ func TestGetSnapshot(t *testing.T) {
 				tk.MustQuery("select * from t for update").Check(testkit.Rows("1", "3"))
 				ts, err = mgr.GetStmtForUpdateTS()
 				require.NoError(t, err)
-				compareSnap2 := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap2 := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err = mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap2, snap))
@@ -396,7 +400,7 @@ func TestGetSnapshot(t *testing.T) {
 			check: func(t *testing.T, sctx sessionctx.Context) {
 				ts, err := mgr.GetStmtReadTS()
 				require.NoError(t, err)
-				compareSnap := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err := mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap, snap))
@@ -406,7 +410,7 @@ func TestGetSnapshot(t *testing.T) {
 				tk.MustQuery("select * from t for update").Check(testkit.Rows("1", "3"))
 				ts, err = mgr.GetStmtForUpdateTS()
 				require.NoError(t, err)
-				compareSnap2 := sessiontxn.GetSnapshotWithTS(sctx, ts)
+				compareSnap2 := internal.GetSnapshotWithTS(sctx, ts)
 				snap, err = mgr.GetSnapshotWithStmtReadTS()
 				require.NoError(t, err)
 				require.True(t, isSnapshotEqual(t, compareSnap2, snap))
@@ -438,6 +442,65 @@ func TestGetSnapshot(t *testing.T) {
 			tk.MustExec("rollback")
 		})
 	}
+}
+
+func TestSnapshotInterceptor(t *testing.T) {
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
+	defer clean()
+
+	tk := testkit.NewTestKit(t, store)
+	tk.MustExec("create temporary table test.tmp1 (id int primary key)")
+	tbl, err := tk.Session().GetDomainInfoSchema().(infoschema.InfoSchema).TableByName(model.NewCIStr("test"), model.NewCIStr("tmp1"))
+	require.NoError(t, err)
+	require.Equal(t, model.TempTableLocal, tbl.Meta().TempTableType)
+	tblID := tbl.Meta().ID
+
+	// prepare a kv pair for temporary table
+	k := append(tablecodec.EncodeTablePrefix(tblID), 1)
+	require.NoError(t, tk.Session().GetSessionVars().TemporaryTableData.SetTableKey(tblID, k, []byte("v1")))
+
+	initTxnFuncs := []func() error{
+		func() error {
+			err := tk.Session().PrepareTxnCtx(context.TODO())
+			if err == nil {
+				err = sessiontxn.GetTxnManager(tk.Session()).AdviseWarmup()
+			}
+			return err
+		},
+		func() error {
+			return sessiontxn.NewTxn(context.Background(), tk.Session())
+		},
+		func() error {
+			return sessiontxn.GetTxnManager(tk.Session()).EnterNewTxn(context.TODO(), &sessiontxn.EnterNewTxnRequest{
+				Type:        sessiontxn.EnterNewTxnWithBeginStmt,
+				StaleReadTS: 0,
+			})
+		},
+	}
+
+	for _, initFunc := range initTxnFuncs {
+		require.NoError(t, initFunc())
+
+		require.NoError(t, sessiontxn.GetTxnManager(tk.Session()).OnStmtStart(context.TODO(), nil))
+		txn, err := tk.Session().Txn(true)
+		require.NoError(t, err)
+
+		val, err := txn.Get(context.Background(), k)
+		require.NoError(t, err)
+		require.Equal(t, []byte("v1"), val)
+
+		val, err = txn.GetSnapshot().Get(context.Background(), k)
+		require.NoError(t, err)
+		require.Equal(t, []byte("v1"), val)
+
+		tk.Session().RollbackTxn(context.Background())
+	}
+
+	// Also check GetSnapshotWithTS
+	snap := internal.GetSnapshotWithTS(tk.Session(), 0)
+	val, err := snap.Get(context.Background(), k)
+	require.NoError(t, err)
+	require.Equal(t, []byte("v1"), val)
 }
 
 func checkBasicActiveTxn(t *testing.T, sctx sessionctx.Context) kv.Transaction {
